Display the map offset for each frame.
The dlopen of a shared library in an apk results in large map offsets.
Unfortunately, the current way that the frame data is printed, it's
impossible to tell what the relative pc is relative to. With the
addition of the offset, it's possible to figure out what the relative
pc actually references.

Bug: 23348999
(cherry picked from commit 802ae55fd1454226b1d76e4a311b2b937d1e0dd9)

Change-Id: I950f92c1cb29ee05eed777f47453efa03318cf3e