unwinder: always set `elf` to null
if `map_info == nullptr`, `elf` will be a garbage value when passed to
`FillInFrame`. In production, this is always null because we zero-init
things, but it's generally best to have it in the code anyway.

Caught by the static analyzer:

system/unwinding/libunwindstack/Unwinder.cpp:221:15: warning: 2nd
function call argument is an uninitialized value
[clang-analyzer-core.CallAndMessage]

Bug: None
Test: TreeHugger
Change-Id: Ia625dd13f0cf01bfcfe26abf6fb4c2b515d38ecb