Fix possible multiple elf objects creation.
If there is a read-only map followed by a read-exec map only
available in memory, and there is an attempt to get the Elf object
from the read-only map, then the read-exec map, there will be two
Elf objects. Since those two objects represent the same elf, this
fixes that so they share the same object.

Test: New unit tests pass.
Change-Id: I8aae149ef7a505591314d38cfabd7a17f6dddfb9