Revert "Clear PAC bits in android_unsafe_frame_pointer_chase()."
This reverts commit 34cd757487578b83f81abd57176e9df65598b97f.

Reason for revert: Breaks ART module: I'm reverting only this change and not the bionic one as it looks to me this change doesn't have any functional change. The reason this breaks the ART module boundary is that ART embeds its own libunwindstack, and that needs to run across different bionic versions (alternatively we could update the bionic version we're currently using as the ART module hasn't shipped yet, but by reverting this CL I want to raise awareness that such changes in libunwindstack are problematic).

Change-Id: If6e3331ac4d80d77c72ddf49b14124e5faa5280f
Exempt-From-Owner-Approval: clean revert of breaking change.