Maps contains shared_ptr MapInfo objects now.
In order to make it easier to change the map data, make the individual
MapInfo objects shared_ptr objects.

This also changes the LocalUpdatableMaps object so that all of the
reparsing is in the object itself. This means this object can be
used with the normal unwinder unchanged and I can delete the
LocalUnwinder completely in a future cl.

Add a method to share a maps object between ThreadUnwinder objects.

Add new unit tests to cover the changes.

Bug: 120606663

Test: Unit tests pass.
Test: Did a performance run before and after to verify no
Test: performance issues.
Change-Id: I074dc925c8e5ff8b3d75779052aefe4801b2c7bf