Fix arm get local registers assembler.
The assembler that grabs the local registers doesn't work so well
if the base register winds up being r1. The code winds up corrupting
memory, and not setting the lr, sp, pc registers because it overwrites
the base register and then writes the three registers to that base
register value.

Test: Ran unit tests.
Change-Id: I953d098280dbadee00ee9f0fb06edd2d209162f7