Allow Unwinder object to be used for local unwinds.
This is the first step to delete the LocalUnwinder object, which does
not operate the same as the Unwinder object. This moves all of the
updatable part of a local unwind into the LocalUpdatableMaps object
so that it is not necessary to have the LocalUnwinder object.

Added unit tests to verify the new use of LocalUpdatableMaps works
when the threads are created after the map data is read.

Test: libbacktrace unit tests, libunwindstack unit tests
Change-Id: I72361d8d2fd9ae7a994c1a391a80d7262bf6e323