Rewrite DwarfSection::InsertFde
The std::map allocations and manipulations are very expensive.
They take more time than the decompression of the debug_frame.

Fortunately, we no longer need the flexibility of the std::map
since the whole debug_frame section is indexed at the start,
and we can move the equivalent functionality there instead.

Bug: 110133331
Test: libunwindstack_unit_test
Test: measure real-world performance changes in perfetto
Change-Id: Ic85f4ffc1b1857089d41ee8bb4f8ca5953fc3ea2