jit/dex info: mask out top byte of 64 bit addresses
The dex info in particular is currently heap-allocated, so the addresses
we read can have a tag in the top-byte if the process we're unwinding
has heap pointer tagging enabled.

Tested locally on an internal master build on a device with heap tagging
enabled, traced_perf's profiles of deskclock no longer have broken
interpreter unwinds.

Bug: 185312982
Test: art/test.py -r -t 137
Change-Id: I8be138514ba60e60280d5c061b5d6ea400da4ed1