Fix race in JitDebug
It is possible that JIT entry is deleted in ART between the
full update and the CheckSeqlock for the returned symbol.
This results in the lookup failing with no more retries.

The CheckSeqlock is still important when returning cached
entries since they might be arbitrarily old.  However,
all entries after full update are guaranteed to be recent,
so we do not need the CheckSeqlock in the first place.

This means the returned symbol might have been just deleted.
However, that is ok, since that is the same as the symbol
being deleted just after we return from the function.

Test: Run ART gcstress tests manually
Change-Id: Ia846008913fc46990092971bb50fcb1921efc301