MapInfo: Lazy allocation of ELF-related fields.
MapInfos are surprisingly big in system-wide profiling.
(currently taking 20% of the whole profiling process,
not even counting the map name strings).

This is because the number of processes times number of
maps per process quickly adds up.

However, many of the fields in MapInfo are related to
ELF caching/loading.  The number of ELF maps is
relatively small compared to grand total of MapInfos.

This CL therefore splits the fields into own struct,
and allocates this struct on first access.  This is
happening transparently under the accessors.

The memory use from the lazy allocated structs is
negligible and this CL as good as removing the fields.
The memory use from MapInfos approximately halves.

Test: libunwindstack_unit_test
Test: compare prefetto system-wide memory usage profiles
Change-Id: I2d1f345c2c0b7d7ffb9d6f815e07e2a63939fd89