Fix multiple_thread_get_elf_fields test.
The test was taking in the i value by reference, which means it was
changing while the thread was being created and caused an abort
with asan enabled.

Also, changed the test slightly so that it only cares about the elf
fields pointer, since all of the fields are 0/nullptr.

Bug: 188452344

Test: Unit tests pass after building with SANITIZER_HOST=address.
Change-Id: I8b0969cbfd43f7e20769e4fc3af65b2f9a47b521
(cherry picked from commit b7328269d7f5fe4407d7267750992a147e041d80)