Move offline test files to offline_files/.
Moved tests/files/offline to a new folder offline_files/. This was
done so that sample offline unwind files are available to both
UnwindOfflineTest.cpp, OfflineUnwindBenchmarks.cpp, and any future
offline unwind related tests/benchmarks.

A new directory utils/ was also added so that unwind offline tests
and benchmarks use the same function to get the string path to the
offline_files/ directory.

Test: Refactoring CL. Benchmarks and unit tests still run/pass.
Change-Id: I4706c2af0fbfb01ec7baabf1814f0f1165a3712f