Allow process memory to be set on UnwindFromPid.
This will allow custom types of process memory objects to be used
to override the normal way of creating process memory.

Add new unit test for the new behavior.

Also, fix a bug where the code would set the process memory to be
the uncached incorrectly modifying the caller's process memory too.

Bug: 189803009

Test: Unit tests pass.
Change-Id: I58e19c6fd8ea2bbf29b10223cf57be3b828968c8