Avoid pathologically high memory use for DEX files.
Do not try to open the DEX file if the file name ends with "(deleted)".
It does not exist. This happens when an app is background-optimized by
ART and all of its files are replaced.

Furthermore, do NOT try to fallback to in-memory copy. It would work,
but all apps tend to be background-optimized at the same time, so it
would lead to excessive memory use during system-wide profiling
(essentially copying all dex files for all apps: hundreds of MBs).

This will cause missing symbols in the backtrace, however,
that outcome is inevitable anyway, since we can not obtain
mini-debug-info for the deleted .oat files.

Bug: 189098772
Bug: 190737786
Test: system-wide profiling while forcing the apps to recompile
Change-Id: I8cefe8eab76dd5020c7ebd4b51a0b89ce17748ef