Move compression memory object into its own file.
The Memory.cpp file is crowded with too much code making it harder
and harder to find modify. Since the MemoryXz object is a pretty
substantial piece of code, break it out into its own file.

Also, remove the extra include files in ElfInterface.cpp that reference
the compression includes since that is all encapsulated in the MemoryXz
object.

Test: Builds and all unit tests pass.
Change-Id: I194c9d1309192e7fe08594af15bf4a52537cffe8