Refactor OfflineUnwindUtils and dependent classes.
This refactor includes the following:
* Change the inheritance relationsips that OfflineUnwindBenchmark and
UnwindOfflineTest have with OfflineUnwindUtils to composition. While
both fixture classes exposed the complete interface of
OfflineUnwindUtils, composition adds to readability and extensibility of
the utils class.
* Create one benchmark fixture method that captures the shared code
between each of the OfflineUnwindBenchmarks. Much of the code between
these benchmarks were very similar so creating a single method reduced
the amount of code by a significant amount.
* This CL also includes some minor refactoring changes such as top-level
comments for each translation unit and using unnamed namespaces to
incorporate internal linkage for benchmarks/tests.

Bug: 192012600
Test: Refactoring CL. Unit tests still pass and benchmarks results
were not altered.

Change-Id: I510e309fff8b43487bae4b93a46abcc8bc283c66