unwind_for_offline: option to perform unwind in specific elf.
This CL adds a command line flag to specify the soname of the ELF that
the user would like the unwind to take place within. Error checking is
done to ensure the provided .so file is actually used by the process and
allows the user to gracefully quit with CTRL-C in case
unwind_for_offline in never able to pause the process inside the
specified ELF.

This option can be used in conjunction with the option to generate/dump
the call stack for all threads of a process. If both options are
selected, unwind_for_offline will wait until at least one thread is
within the desired ELF. At that point, all threads will be unwound
in their current state.

This feature will be useful for obtaining realistic/representative
unwinds for profiler-like offline unwind benchmarks. For example, this
option may be used to help ensure different unwinds within the same
process use a different set of ELF files. cferris also indicated that
this feature would be very useful when it comes to testing and
debugging unwinds with specific shared libraries. For example, one
OEM failed to perform an unwind with a specific shared library. The
ideal method to reproduct the issue would have been to run
unwind_for_offline with that ELF file selected.

Bug: 192012600
Test: Manual testing. Test cases:
* Incorrect soname or ELF file that process does not use.
* soname used by the process but we are unable to pause the process
within that ELF (never reached) --> graceful exit with CTRL-C. Used
libutil.so with Android bluetooth process.
* PC lies within specified ELF upon calling unwind_for_offline(e.g.
libc.so for most processes).
* PC does not initially lie within specific ELF but eventually lies
within ELF and successfully unwinds. Targetted libbluetooth.so and
blueooth process. Started at home screen then tried to pair with
bluetooth headphones to trigger the unwind.
* Verify unwind within specific ELF option works successfully with and
without dump all threads option (-t).

Change-Id: Ib7c6c0981e2aaaef1f08c24250d76160f7c2b6e4