OfflineUnwindUtils: Support consecutive unwinds.
This CL refactors OfflineUnwindUtils to allow clients to perform
multiple/consecutive unwinds. In the previous version of
OfflineUnwindUtils, a client would need to create multiple
OfflineUnwindUtils instances in order to test or benchmark offline
unwinds. This would require modifying the existing test/benchmark fixtures
and/or each of the individual test/benchmarks to support this change.
This change simply adds a new Init overload for consecutive offline
unwinds and leaves the remainder of the public API the same.

Consecutive unwinds are necessary for profiler-like offline unwind
benchmarks. See b/192012600 for more for more information regarding
these benchmarks.

Bug: 192012600
Test: Refactoring CL. Benchmarks still run and unit tests still pass.
Change-Id: I5ebb8606eaf4bf144690de41283619218108e3b8