unwind_for_offline: option to dump backtraces for all threads.
Currently, offline_for_unwind can only generate offline unwind files
for the main thread of the selected process. This change adds a
command line flag to specify whether or not the user wants to
gather and dump offline unwind files for all threads of a process.

This feature will be useful for obtaining unwinds for both the
single-process profiling scenario (one process, multiple thread case)
for profiler-like offline unwind benchmarks as well as general
ability to obtain offline unwinds from a specific threads.

This CL also includes the implementation for ProcessTracer, a class that
abstracts operations for tracing a process and its threads with
ptrace(2).
The original idea to create this class was to handle graceful detatching
and resuming of the managed process if a user chooses to stop waiting
for the process to enter the desired ELF in CL 1769946. Additionally,
after a discussion with cferris, I learned that other libunwinstack
tools and unit tests would also greatly benefit from having a signle
interface to handle tracing processes with ptrace.

Bug: 192012600
Test: Manually testing to verify unwinds were successful for single
threaded and multithreaded processes with and without the new 'dump all
threads' flag.

Change-Id: Ic646d830151305a1906467c175b04912b91dc98c