Add ELFs for large ELF benchmarks.
Add the identified ELF files needed for the large ELF file benchmarks.
See b/192012600 for more for more information regarding these
benchmarks.

I attempted to stripping the text and un-needed debug_* sections out of
each of the ELFs with objcopy and manual modification of the ELF header
(e_machine field in header was deleted for some reason). This appears to
work for the ELF creation benchmarks (ElfBenchmark::BenchmarkElfCreate).
However I ran into issues for the symbolization benchmarks
(SymbolBenchmarks.cpp) for libLLVM_android.so. So, in the interest of
time (approved by cferris) the stripped version of libLLVM_android.so
will only be used in the ELF creation benchmarks whereas libpac.so
will be used in both benchmark collections.

Bug: 192012600
Test: N/A. Only uploading benchmarking target files.
Change-Id: I978b2bc5f4732f5a6ee0fa52663a60081481002e