Remove the error message for unsupported elf.
Originally, this message helped to diagnosis some early bugs
in libunwindstack. But now this message is not every useful.
It can sometimes trigger when trying to read elf files that
represent JIT'd code when there is a race between the elf
being in written or removed. There are many different ways
that an elf can be malformed, or unsupported, but only this
one particular method logs an error. Since that doesn't
really match the rest of code, simply remove this one error
message.

Bug: 198465745

Test: Unit tests pass.
Change-Id: Ic63c1c0f0c9383579b7d2e896079e177575044c8