Allow a global to be in a zero offset rw map.
It's possible if an elf file is smaller than a page size, then
the global variable will be in a zero offset rw map. This change
removed the check to verify that it was in a non-zero offset
map.

Add new unit tests for the Global object to cover the new case
and other tests cases.

Bug: 199043576

Test: Verified a zero offset rw section with debuggerd tests.
Test: All new unit tests pass.
Change-Id: I6af9a70da52c27555b8ed45dc4511f020aaba9d3