Fix missing offset in display for apk backtrace.
When an apk maps a shared library that only has a single read-executable
map and the original apk files is unreadable, the offset data is
not set properly. Specifically, this is the output before the bug:

  #01 pc 000000000222675c  GoogleCamera.apk

After the fix, the line now looks like so:

  #01 pc 000000000222675c  GoogleCamera.apk (offset 0x269f000)

There is no problem if the shared library creates a read-only,
then read-executable map.

Added a new unit test to cover this case.

Bug: 201703079

Test: All unit tests pass.
Test: Also, created an offline test to verify that with this change
Test: the offset is displayed properly. The offline test will be
Test: added in a follow-up cl.
Change-Id: I3cc840beb6b3005afade06a4d0ae400e559a0046