Add offline test for apk issues.
This cl adds a few small test cases related to apks. Specifically,
adding two cases where an apk is only available in memory and
verifying that the unwind does display offset data. In addition,
use the same two tests cases, but allow the apk to be readable.

One unreadable test case covers the problem found in the given
bug, where an ureadable map out of an apk doesn't print the
offset data if there is only a single map containing the elf
data and the executable data.

Bug: 201703079

Test: All new unit tests pass.
Change-Id: Ib8409c84e5c18928f34d66b16f7fa0d50e8b4300