Remove the next_real_map and prev_real_map.
It's a pain to keep track of these pointers when doing a reparse.
Rather than keeping these in sync, remove them completely and add
a next_map. The next_real_map and prev_real_map variables are only
needed in very specific circumstances that are very unlikely to
occur. In those cases, derive the values when they are needed.

Also, make the Get{Next,Prev}RealMap functions verify that the
name is the same so that those checks can be removed from the
places that use those values.

Add new unit tests to verify the new behavior, and verify the
pointers are set properly.

Test: All unit tests pass.
Change-Id: I8fd784df71d7b39890c763755bec5904c0bed8a3