Move MapInfo objects in to frame data.
The frame data when doing an unwind contains a lot of map_XXX
fields which were copies of the data from a MapInfo object.
Instead of doing these copies, replace those fields with a
shared pointer of the MapInfo object itself.

Remove the SetEmbeddedSoname mechanism. Since the MapInfo object
is in the frame data, move this functionality into a MapInfo
function called GetFullName() that will return the map name plus
the soname if needed. A caller needs to call this function directly,
so there is no need to avoid getting this data during the unwind.

Create a GetLoadBias() function that does not require process_memory.
This new function will work for any MapInfo object that appears
in frame data.

Adjust tests for these new changes.

Bug: 120606663

Test: Unit tests pass.
Change-Id: I59199aa344e4b44c158f2636f7f7beff53d9b99a