Add detailed error messages for remote unwinds.
If the RemoteGet or RemoteGetArch fail, they don't properly indicate
why.  When using the AndroidRemoteUnwinder object, it can fail and
it can be unclear why. So add error messages that get logged and a
different error code to be returned by unwind calls.

Add unit tests for the ptrace failing, but there is no way to add a
test to verify if the data structures are different. For that case,
I hand modified the code and verified the unsupported error code
was returned.

Test: New unit tests pass.
Test: Modify the code so that the ptrace error is ignored and the secondary
Test: failure is executed.
Change-Id: Icd5a6b942f65bf97eed7565c88c2dd5d6c5b2d7d