Modify the max size algorithm.
Some shared libraries seem to be putting a dynamic symbol table
at the end of the shared library. Modify the GetMaxSize()
function to read the PT_LOAD values and if that is larger
than the computed elf size, use that value instead.

Added new offline test to expose this problem.

Bug: 268566992

Test: All unit tests pass.
Change-Id: I9c240e83d6090417fc1f642f049175def908be32