Fix race condition in tests.
A few UnwindTest thread tests can fail because the thread to be
unwound has only just called OuterFunction(). The test assumes
that the thread has finished calling all of the way down to the
InnerFunction(). Modify these tests to actually wait until
InnerFunction is called.

Remove redundant calls to ResetGlobals() that is already
done in the SetUp() function.

In addition, fix a race in AndroidLocalUnwinderTest that fails
because the test expects a function at the top of the stack.
Allow that function to be in an any position instead.

Bug: 273963946

Test: Ran a lot.
Change-Id: Id00fece3a6ce1f08a821f9e0e2572b1390a22537