libunwindstack: Fix AsmGetRegs to not assume Thumb state
Arm implementation of RegsGetLocal unconditionally switches to the ARM
state ("bx pc") and then back to Thumb. This works fine in Thumb code,
but breaks if called from ARM code (for example, if libunwindstack
is built with -marm): it leaves the CPU in Thumb state, executing A32
instructions.

This patch adds ifdefs so that state changes only happen if the code is
compiled in Thumb mode. Note that this fix isn't comprehensive[1], but
it's simple and avoids adding a trampoline and having to stash the LR.

[1] For example, it doesn't handle the theoretical case where this code
is called from a function with the target("arm") attribute in a module
compiled with -mthumb.

Test: UnwindTest.local w/ instruction_set: "arm" in Android.bp
Change-Id: I9f08c58071a7353967fd447e4c0e86cbb09da2f9