libunwindstack uses shared libz instead of static libz
libz is part of the ndk and besides some exceptions (e.g. linker),
should not be linked statically. libunwindstack has multiple variants in
the build (platform, runtime, art) and changing the link to shared also
ensures that we can enforce an api boundary between libz and
libunwindstack based on the specific variant.

Test: presubmits
Bug: 323067797
Bug: 281077552
Change-Id: Iade91546d1f8e4fa01e1ddbe8081720aec590d6c