Add special processing of linker names.
The linker has names of the form __dl_XXX where XXX can be a mangled
name. Recognize this special case and demangle the XXX names so that
the function names are easier to understand. The new name will be
__dl_YYY where YYY is the demangled name.

Fix names in a few offline unwinds to match the new linker demangles.

Bug: 360247782

Test: New unit tests pass.
Change-Id: I2846cdbce16c0ee4ec2f54f1fc35778d8040d95e