Fix failure checking undefined register on bionic.
When verifying that the return address in bionic is undefined
in __libc_init, the code did not pass an arch value when getting
the dwarf location. Normally this doesn't matter but a partner
has dwarf data that includes the arm64 pseudo register. This
means that processing fails because the pseudo register only
exists for arm64. Now pass the arch around properly.

Refactor the test a bit and use the elf object in the
map info object instead of trying to recreate it. Also,
check gnu_debugdata (debug_frame/eh_frame) for the return
address if not found in the regular debug_frame or eh_frame.

Bug: 417155681

Test: All unit tests pass on arm/arm64/x86/x86_64.
Change-Id: If49c2972e23703663bc4ca5eca223f82c542cb42