Small cleanup of arm64 registers/structures.
There is no reason to have the arm64 user structure match the
exact naming in the kernel structures. Instead make it an array
for the whole set of registers so they can all be copied in a
single memcpy operation.

Also, avoid using ARM64_REG_LAST and use exactly what the registers
are. This is because it's likely we will need to add support
for extra registers similar to riscv vlenb. It also makes it
clear how many registers we are expecting.

Fix extra pstate in UcontextArm64.h that is not needed. This was
found by user csanad.hajdu@arm.com.

Test: All unit tests pass.
Change-Id: I20b4ac6e2ac19a3992685af334a24f52eaff85f3