Move extra registers to list of regular registers.
Instead of having a separate set of extra registers, just add the
extra registers to the list of real registers.

This removes a lot of special case code in debuggerd and in the
unwinder.

Update tests for the new extra registers code handling.

Change the default convert function to return the max register
to cause an error if an unknown register is mentioned in the
dwarf unwind info.
Fix some issues where registers should be converted and weren't.

Test: All libunwindstack/debuggerd unit tests pass.
Change-Id: I5a8895a466db71f61a48f083565a2027b534b5ab