Fix a few issues with VG register support.
Assume the vl value retrieved through ptrace is valid if the returned
size is greater than the header size. Some kernels return a smaller
than header size value if the vl is not valid.

Check independently for SVE and SSVE since it's possible to have
either one enabled and the other disabled.

Test: All unit tests pass on arm64 device.
Change-Id: Ief8704bdf8bf5ea25e8dc146486f32276107d457