Standardize the way to get extra registers.
Add a public GetVgFromRemote static function to RegsArm64 to be
similar to the RegsRiscv64 GetVlenbFromRemote.

Change the default read value for arm64/riscv64 to -1 since 0
is techincally a valid value. Update the tests to match.

From the Read functions that take a pid, avoid any remote read
calls if pid is -1. The simpleperf code calls the read call without
a pid in offline cases, so we don't want to try and do ptrace calls
that are guaranteed to fail.

Test: All unit tests pass.
Change-Id: Id2d810ecec54f44712c9cfe25847698c8017450c