Use OWNERS rather than trendy teams.
Several of these are wrong, none of these were useful to us, and -- after having been told to waste time adding this stuff a year or two ago -- we're now being told to waste time removing this stuff again, and going back to the thing we said should be used in the first place. Progress!

Bug: http://b/416765999
Change-Id: Iffa2b06529eaefec8fc46c8b8108c4bcddadf35d